Move spending plans to their own crate
